pyproject.toml
[tool.poetry]
name = "polyflexmd"
version = "0.1.0"
description = ""
authors = ["Yahor Paromau <yahor.paromau@gmail.com>"]
readme = "README.md"
packages = [{include = "polyflexmd", from="./src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
typer = "^0.7.0"
numpy = "^1.24.2"
spock-config = "^3.0.2"
pydantic = "^1.10.7"
tomlkit = "^0.11.7"
gitpython = "^3.1.31"
jinja2 = "^3.1.2"
notebook = "^6.5.3"
pandas = "^1.5.3"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
pymatgen = "^2023.3.23"
scipy = "^1.10.1"
jupyter = "^1.0.0"
pandarallel = "^1.6.4"
nbconvert = "^7.3.1"
papermill = "^2.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
